+++ kotlin/src/main/kotlin/org/jetbrains/research/testspark/kotlin/KotlinPsiHelper.kt
@@ -160,9 +160,9 @@ class KotlinPsiHelper(var psiFile: PsiFile) : PsiHelper {
         val ktFunction = getSurroundingMethod(caret.offset)
         val line: Int? = getSurroundingLine(caret.offset)?.plus(1)
 
-        ktClass?.let { result.add(getClassDisplayName(it)) }
-        ktFunction?.let { result.add(getMethodDisplayName(it)) }
-        line?.let { result.add(getLineDisplayName(it)) }
+        ktClass?.let { result.add(getClassHTMLDisplayName(it)) }
+        ktFunction?.let { result.add(getMethodHTMLDisplayName(it)) }
+        line?.let { result.add(getLineHTMLDisplayName(it)) }
 
         if (ktClass != null && ktFunction != null) {
             log.info(
@@ -176,12 +176,12 @@ class KotlinPsiHelper(var psiFile: PsiFile) : PsiHelper {
         return result.toArray()
     }
 
-    override fun getLineDisplayName(line: Int) = "<html><b><font color='orange'>line</font> $line</b></html>"
+    override fun getLineHTMLDisplayName(line: Int) = "<html><b><font color='orange'>line</font> $line</b></html>"
 
-    override fun getClassDisplayName(psiClass: PsiClassWrapper): String =
+    override fun getClassHTMLDisplayName(psiClass: PsiClassWrapper): String =
         "<html><b><font color='orange'>${psiClass.classType.representation}</font> ${psiClass.qualifiedName}</b></html>"
 
-    override fun getMethodDisplayName(psiMethod: PsiMethodWrapper): String {
+    override fun getMethodHTMLDisplayName(psiMethod: PsiMethodWrapper): String {
         psiMethod as KotlinPsiMethodWrapper
         return when {
             psiMethod.isTopLevelFunction -> "<html><b><font color='orange'>top-level function</font></b></html>"
